add validations habilidad

+++ src/main/java/com/apirest/portfolio/controller/HabilidadController.java
@@ -4,9 +4,11 @@
  */
 package com.apirest.portfolio.controller;
 
+import com.apirest.portfolio.dto.Mensaje;
 import com.apirest.portfolio.model.Habilidad;
 import com.apirest.portfolio.service.IHabilidadService;
 import java.util.List;
+import javax.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -31,43 +33,61 @@ public class HabilidadController {
     @Autowired
     private IHabilidadService interHabilidad;
     
+    /**
+     *
+     * @return
+     */
     @GetMapping("habilidad/traer")
     public ResponseEntity<List<Habilidad>> getHabilidades(){
         try {
             List<Habilidad> hab = interHabilidad.getHabilidades();
             if (hab.isEmpty()){
-                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+                return new ResponseEntity(new Mensaje("No hay habilidades"), HttpStatus.NO_CONTENT);
             }
             else {
                 return new ResponseEntity<>(hab, HttpStatus.OK);
             }
         } catch (Exception e){
-            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
+            return new ResponseEntity(new Mensaje(e.getMessage()), HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
     
+    /**
+     *
+     * @param hab
+     * @return
+     */
     @PostMapping("habilidad/crear")
-    public ResponseEntity<Habilidad> createHabilidad(@RequestBody Habilidad hab){
+    public ResponseEntity<Habilidad> createHabilidad(@Valid @RequestBody Habilidad hab){
         try {
             interHabilidad.saveHabilidad(hab);
             return new ResponseEntity<>(hab, HttpStatus.OK);
         } catch (Exception e){
-            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
+            return new ResponseEntity(new Mensaje(e.getMessage()), HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
     
+    /**
+     *
+     * @param id
+     * @return
+     */
     @DeleteMapping("habilidad/borrar/{id}")
     public ResponseEntity<Habilidad> deleteHabilidad(@PathVariable Long id){
         try {
-            Habilidad hab = interHabilidad.findHabilidad(id);
-            interHabilidad.deleteHabilidad(id);
-            return new ResponseEntity<>(hab, HttpStatus.OK);
+            if (interHabilidad.existHabilidadById(id)){
+                Habilidad hab = interHabilidad.findHabilidad(id);
+                interHabilidad.deleteHabilidad(id);
+                return new ResponseEntity<>(hab, HttpStatus.OK);
+            } else {
+                return new ResponseEntity(new Mensaje("No se borro nada porque no existe un registro con ese id"), HttpStatus.OK);
+            }
         } catch (Exception e){
-            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
+            return new ResponseEntity(new Mensaje(e.getMessage()), HttpStatus.INTERNAL_SERVER_ERROR);
         }  
     }
     
-    @PutMapping("habilidad/editar/{id}")
+    /*@PutMapping("habilidad/editar/{id}")
     public ResponseEntity<Habilidad> editHabilidad(
             @PathVariable Long id,
             @RequestParam("descripcion") String descripcion,
@@ -81,24 +101,59 @@ public class HabilidadController {
             interHabilidad.saveHabilidad(hab);
             return new ResponseEntity<>(hab, HttpStatus.OK);
         } catch (Exception e){
-            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
+            return new ResponseEntity(new Mensaje(""), HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }*/
+
+    /**
+     *
+     * @param id
+     * @param hab
+     * @return
+     */
+    @PutMapping("habilidad/editar/{id}")
+    public ResponseEntity<Habilidad> editHabilidad(
+            @PathVariable Long id,
+            @Valid @RequestBody Habilidad hab){
+        try {
+            if (interHabilidad.existHabilidadById(id)){
+                Habilidad habilidad = interHabilidad.findHabilidad(id);
+                habilidad.setDescripcion(hab.getDescripcion());
+                habilidad.setNivel(hab.getNivel());
+                habilidad.setTipo_habilidad_id(hab.getTipo_habilidad_id());
+                interHabilidad.saveHabilidad(habilidad);
+                return new ResponseEntity<>(habilidad, HttpStatus.OK);
+            } else {
+                return new ResponseEntity(new Mensaje("No existen registros para el id: " + id), HttpStatus.NOT_FOUND);
+            }
+        } catch (Exception e){
+            return new ResponseEntity(new Mensaje(e.getMessage()), HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
     
     //hecho por mi y ya esta funcionando
-    @GetMapping("habilidad/traerById/{id}")
+    /*@GetMapping("habilidad/traerById/{id}")
     public List<Habilidad> findHabilidadByUserId(@PathVariable Long id){
         List<Habilidad> lista = interHabilidad.getHabilidadesByUserId(id);
         return lista;
-    }
-    
+    }*/
+
+    /**
+     *
+     * @param usuario
+     * @return
+     */
     @GetMapping("habilidad/buscarByUsuario/{usuario}")
     public ResponseEntity<List<Habilidad>> bucarByUsuario(@PathVariable("usuario") String usuario){
         try{
             List<Habilidad> listaHabilidad = interHabilidad.getHabilidadesByUsuario(usuario);
-            return new ResponseEntity<List<Habilidad>>(listaHabilidad, HttpStatus.OK);
+            if (listaHabilidad.isEmpty()){
+                return new ResponseEntity(new Mensaje("No hay registros para el usuario: " + usuario), HttpStatus.NOT_FOUND);
+            } else {
+                return new ResponseEntity<>(listaHabilidad, HttpStatus.OK);
+            }
         } catch (Exception e){
-            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
+            return new ResponseEntity(new Mensaje(e.getMessage()), HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 }
